requirements.txt
imutils==0.5.4
numpy==1.26.2
opencv-python==4.8.1.78
paho-mqtt==1.6.1
pyserial==3.5
PyYAML==6.0.1
